Multiple Oscillators!
Didn't want to just repeat all the code so tried creating new osc objects but that wouldn't allow for separate ADSRs, only unison(ish) stuff. creating new apvts involved auto generating variable names (bad idea), arrays of references don't work and casting to int/float just gave the memory address, this is the correct solution - just need them all to play at the same time now.

+++ OscData.cpp
@@ -12,7 +12,7 @@
 
 void OscData::setWaveType(const int waveType) {
 
-    switch (waveType)
+    switch ((int)waveType)
     {
     case 0:
         initialise([](float x) {return std::sin(x); });
